requirements.txt
fastapi
uvicorn
PyPDF2
langchain
sentence-transformers
google-generativeai
langchain-community
langchain-google-genai
python-dotenv
jinja2
transformers
torch
pydantic
pdfplumber
deep-translator
faiss-cpu
sentencepiece
python-multipart
tensorflow

